Did some refactoring work. Templates!
Moved Socks things into a new folder, and replaced concrete references
to Sockets with template parameters. Added the first step towards
polymorphic state-based request handling in the SocksSession, which is
implemented by extensions of the interface ISocksStateManager (this
should probably be re-named ISocksState, since the Session is
responsible for managing it).